refactor: extract provider operations status helpers
